TASK: Move `<I18n/>`-component into separate module
