+++ packages/neos-ui-i18n/package.json
@@ -3,7 +3,7 @@
   "version": "",
   "description": "I18n utilities and components for Neos CMS UI.",
   "private": true,
-  "main": "./src/index.tsx",
+  "main": "./src/index.ts",
   "devDependencies": {
     "@neos-project/jest-preset-neos-ui": "workspace:*",
     "enzyme": "^3.8.0",
