+++ packages/neos-ui-i18n/src/component/I18n.tsx
@@ -1,11 +1,15 @@
+/*
+ * This file is part of the Neos.Neos.Ui package.
+ *
+ * (c) Contributors of the Neos Project - www.neos.io
+ *
+ * This package is Open Source Software. For the full copyright and license
+ * information, please view the LICENSE file which was distributed with this
+ * source code.
+ */
 import React from 'react';
-import {LegacyParameters, i18nRegistry} from './registry';
 
-export {initializeI18n, setupI18n, teardownI18n} from './global';
-
-export type {I18nRegistry} from './registry';
-
-export {translate} from './translate';
+import {LegacyParameters, i18nRegistry} from '../registry';
 
 interface I18nProps {
     // Fallback key which gets rendered once the i18n service doesn't return a translation.
@@ -25,7 +29,7 @@ interface I18nProps {
     className?: string;
 }
 
-export default class I18n extends React.PureComponent<I18nProps> {
+export class I18n extends React.PureComponent<I18nProps> {
     public render(): JSX.Element {
         const {packageKey, sourceName, params, id, fallback} = this.props;
